+++ include/math/Vector2.h
@@ -6,15 +6,15 @@
 
 class Vector2 {
 public:
-    float x, y;
+    double x, y;
 
     Vector2() : x(0.0f), y(0.0f) {}
-    Vector2(float x, float y) : x(x), y(y) {}
+    Vector2(double x, double y) : x(x), y(y) {}
 
-    float getX() const { return x; }
-    float getY() const { return y; }
+    double getX() const { return x; }
+    double getY() const { return y; }
 
-    void set(float x, float y) { this->x = x; this->y = y; }
+    void set(double x, double y) { this->x = x; this->y = y; }
 
     Vector2 operator+(const Vector2& other) const {
         return Vector2(x + other.x, y + other.y);
@@ -24,30 +24,30 @@ public:
         return Vector2(x - other.x, y - other.y);
     }
 
-    Vector2 operator*(float scalar) const {
+    Vector2 operator*(double scalar) const {
         return Vector2(x * scalar, y * scalar);
     }
 
-    Vector2 operator/(float scalar) const {
+    Vector2 operator/(double scalar) const {
         return Vector2(x / scalar, y / scalar);
     }
 
-    float dot(const Vector2& other) const {
+    double dot(const Vector2& other) const {
         return x * other.x + y * other.y;
     }
 
-    float length() const {
+    double length() const {
         return std::sqrt(x * x + y * y);
     }
 
     // Squared length (no square root for optimization)
-    float lengthSquared() const {
+    double lengthSquared() const {
         return x * x + y * y;
     }
 
     // Normalize the vector (make it unit length)
     void normalize() {
-        float len = length();
+        double len = length();
         if (len > 0.0f) {
             x /= len;
             y /= len;
@@ -56,7 +56,7 @@ public:
 
     // Return a normalized copy of the vector
     Vector2 normalized() const {
-        float len = length();
+        double len = length();
         if (len > 0.0f) {
             return Vector2(x / len, y / len);
         }
