+++ src/core/objects/3D/Object3.cpp
@@ -1,6 +1,6 @@
 #include "include/core/objects/3D/Object3.h"
 
-Object3::Object3(Vector3 pos, Vector3 vel, float m) {
+Object3::Object3(Vector3 pos, Vector3 vel, double m) {
     velocity = vel;
     position = pos;
     mass = m;
